move Distribution enum into new submodule mkosi.distributions
