+++ tests/test_backend.py
@@ -4,14 +4,7 @@ import os
 from pathlib import Path
 
 import mkosi.backend as backend
-from mkosi.backend import Distribution, PackageType, set_umask
-
-
-def test_distribution() -> None:
-    assert Distribution.fedora.package_type == PackageType.rpm
-    assert Distribution.fedora is Distribution.fedora
-    assert Distribution.fedora.package_type is not Distribution.debian.package_type
-    assert str(Distribution.fedora) == "fedora"
+from mkosi.backend import set_umask
 
 
 def test_set_umask() -> None:
