+++ tests/test_distributions.py
@@ -1,0 +1,10 @@
+# SPDX-License-Identifier: LGPL-2.1+
+
+from mkosi.distributions import Distribution, PackageType
+
+
+def test_distribution() -> None:
+    assert Distribution.fedora.package_type == PackageType.rpm
+    assert Distribution.fedora is Distribution.fedora
+    assert Distribution.fedora.package_type is not Distribution.debian.package_type
+    assert str(Distribution.fedora) == "fedora"
